HW15: Refactor (move inter Carriage to outside, create Demo)
