--- src/com/company/Main.java
@@ -1,8 +1,0 @@
-package com.company;
-
-public class Main {
-
-    public static void main(String[] args) {
-	// write your code here
-    }
-}
